Replace HostFs completely
