+++ lib/virtual-fs/src/fs/host.rs
@@ -1,3 +1,4 @@
+use crate::ops::PathClean;
 use crate::{
     DirEntry, FileType, FsError, Metadata, OpenOptions, OpenOptionsConfig, ReadDir, Result,
     VirtualFile,
@@ -20,37 +21,31 @@ use tokio::runtime::Handle;
 
 #[derive(Debug, Clone)]
 #[cfg_attr(feature = "enable-serde", derive(Serialize, Deserialize))]
-pub struct FileSystem(Handle);
-
-impl Default for FileSystem {
-    fn default() -> Self {
-        Self(Handle::current())
-    }
-}
-impl FileSystem {
-    pub fn new(handle: Handle) -> Self {
-        FileSystem(handle)
-    }
+pub struct FileSystem {
+    handle: Handle,
+    base: PathBuf,
 }
 
 impl FileSystem {
-    pub fn canonicalize(&self, path: &Path) -> Result<PathBuf> {
-        if !path.exists() {
-            return Err(FsError::InvalidInput);
-        }
-        fs::canonicalize(path).map_err(Into::into)
+    pub fn new(base: PathBuf) -> Result<Self> {
+        Ok(FileSystem {
+            handle: Handle::current(),
+            base: base.canonicalize()?,
+        })
     }
 }
 
 impl crate::FileSystem for FileSystem {
     fn read_dir(&self, path: &Path) -> Result<ReadDir> {
+        let mut base = self.base.clone();
+        let path = self.base.join(path.clean_safely()?);
         let read_dir = fs::read_dir(path)?;
         let mut data = read_dir
             .map(|entry| {
                 let entry = entry?;
                 let metadata = entry.metadata()?;
                 Ok(DirEntry {
-                    path: entry.path(),
+                    path: entry.path().strip_prefix(&mut base).unwrap().to_owned(),
                     metadata: Ok(metadata.try_into()?),
                 })
             })
@@ -61,35 +56,45 @@ impl crate::FileSystem for FileSystem {
     }
 
     fn create_dir(&self, path: &Path) -> Result<()> {
-        if path.parent().is_none() {
-            return Err(FsError::BaseNotDirectory);
+        let path: PathBuf = self.base.join(path.clean_safely()?);
+        if path == PathBuf::from(".") {
+            // Creating the root should always work
+            return Ok(());
         }
         fs::create_dir(path).map_err(Into::into)
     }
 
     fn remove_dir(&self, path: &Path) -> Result<()> {
-        if path.parent().is_none() {
+        let path: PathBuf = self.base.join(path.clean_safely()?);
+        if path == PathBuf::from(".") {
             return Err(FsError::BaseNotDirectory);
         }
         // https://github.com/rust-lang/rust/issues/86442
         // DirectoryNotEmpty is not implemented consistently
-        if path.is_dir() && self.read_dir(path).map(|s| !s.is_empty()).unwrap_or(false) {
+        if path.is_dir()
+            && fs::read_dir(&path)
+                .map(|mut s| s.next().is_some())
+                .unwrap_or(false)
+        {
             return Err(FsError::DirectoryNotEmpty);
         }
         fs::remove_dir(path).map_err(Into::into)
     }
 
     fn rename<'a>(&'a self, from: &'a Path, to: &'a Path) -> BoxFuture<'a, Result<()>> {
-        let from = from.to_owned();
-        let to = to.to_owned();
         Box::pin(async move {
+            let from = from.to_owned().clean_safely()?;
+            let to = to.to_owned().clean_safely()?;
             use filetime::{set_file_mtime, FileTime};
-            if from.parent().is_none() {
+            if from == PathBuf::from(".") {
                 return Err(FsError::BaseNotDirectory);
             }
-            if to.parent().is_none() {
+            if to == PathBuf::from(".") {
                 return Err(FsError::BaseNotDirectory);
             }
+            let from: PathBuf = self.base.join(&from);
+            let to: PathBuf = self.base.join(&to);
+
             if !from.exists() {
                 return Err(FsError::EntryNotFound);
             }
@@ -129,7 +134,8 @@ impl crate::FileSystem for FileSystem {
     }
 
     fn remove_file(&self, path: &Path) -> Result<()> {
-        if path.parent().is_none() {
+        let path: PathBuf = self.base.join(path.clean_safely()?);
+        if path == PathBuf::from(".") {
             return Err(FsError::BaseNotDirectory);
         }
         fs::remove_file(path).map_err(Into::into)
@@ -140,6 +146,7 @@ impl crate::FileSystem for FileSystem {
     }
 
     fn metadata(&self, path: &Path) -> Result<Metadata> {
+        let path: PathBuf = self.base.join(path.clean_safely()?);
         fs::metadata(path)
             .and_then(TryInto::try_into)
             .map_err(Into::into)
@@ -210,6 +217,8 @@ impl crate::FileOpener for FileSystem {
         path: &Path,
         conf: &OpenOptionsConfig,
     ) -> Result<Box<dyn VirtualFile + Send + Sync + 'static>> {
+        let path = self.base.join(path.clean_safely()?);
+
         // TODO: handle create implying write, etc.
         let read = conf.read();
         let write = conf.write();
@@ -221,11 +230,11 @@ impl crate::FileOpener for FileSystem {
             .create(conf.create())
             .append(conf.append())
             .truncate(conf.truncate())
-            .open(path)
+            .open(&path)
             .map_err(Into::into)
             .map(|file| {
                 Box::new(File::new(
-                    self.0.clone(),
+                    self.handle.clone(),
                     file,
                     path.to_owned(),
                     read,
@@ -920,22 +929,26 @@ impl VirtualFile for Stdin {
 mod tests {
     use tempfile::TempDir;
 
-    use crate::host_fs::FileSystem;
+    use super::FileSystem;
     use crate::FileSystem as FileSystemTrait;
     use crate::FsError;
     use std::path::Path;
+    use std::path::PathBuf;
 
     #[tokio::test]
     async fn test_new_filesystem() {
         let temp = TempDir::new().unwrap();
         std::fs::write(temp.path().join("foo2.txt"), b"").unwrap();
 
-        let fs = FileSystem::default();
-        assert!(fs.read_dir(Path::new("/")).is_ok(), "hostfs can read root");
+        let fs = FileSystem::new(temp.path().into()).expect("get filesystem");
+        assert!(
+            fs.read_dir(Path::new("/")).is_ok(),
+            "NativeFS can read root"
+        );
         assert!(
             fs.new_open_options()
                 .read(true)
-                .open(temp.path().join("foo2.txt"))
+                .open(Path::new("/foo2.txt"))
                 .is_ok(),
             "created foo2.txt"
         );
@@ -943,17 +956,17 @@ mod tests {
 
     #[tokio::test]
     async fn test_create_dir() {
-        let temp = TempDir::new().unwrap();
-        let fs = FileSystem::default();
+        let temp: TempDir = TempDir::new().unwrap();
+        let fs = FileSystem::new(temp.path().into()).expect("get filesystem");
 
         assert_eq!(
-            fs.create_dir(Path::new("/")),
-            Err(FsError::BaseNotDirectory),
-            "creating a directory that has no parent",
+            fs.create_dir(Path::new("../")),
+            Err(FsError::InvalidInput),
+            "creating a directory out of bounds",
         );
 
         assert_eq!(
-            fs.create_dir(&temp.path().join("foo")),
+            fs.create_dir(Path::new("/foo")),
             Ok(()),
             "creating a directory",
         );
@@ -963,7 +976,7 @@ mod tests {
             "foo dir exists in host_fs"
         );
 
-        let cur_dir = read_dir_names(&fs, temp.path());
+        let cur_dir = read_dir_names(&fs, "/");
 
         if !cur_dir.contains(&"foo".to_string()) {
             panic!("cur_dir does not contain foo: {:#?}", cur_dir);
@@ -975,7 +988,7 @@ mod tests {
         );
 
         assert_eq!(
-            fs.create_dir(&temp.path().join("foo/bar")),
+            fs.create_dir(Path::new("foo/bar")),
             Ok(()),
             "creating a sub-directory",
         );
@@ -985,14 +998,14 @@ mod tests {
             "foo dir exists in host_fs"
         );
 
-        let foo_dir = read_dir_names(&fs, temp.path().join("foo"));
+        let foo_dir = read_dir_names(&fs, Path::new("/foo"));
 
         assert!(
             foo_dir.contains(&"bar".to_string()),
             "the foo directory is updated and well-defined"
         );
 
-        let bar_dir = read_dir_names(&fs, temp.path().join("foo/bar"));
+        let bar_dir = read_dir_names(&fs, Path::new("/foo/bar"));
 
         assert!(
             bar_dir.is_empty(),
@@ -1002,8 +1015,8 @@ mod tests {
 
     #[tokio::test]
     async fn test_remove_dir() {
-        let temp = TempDir::new().unwrap();
-        let fs = FileSystem::default();
+        let temp: TempDir = TempDir::new().unwrap();
+        let fs = FileSystem::new(temp.path().into()).expect("get filesystem");
 
         assert_eq!(
             fs.remove_dir(Path::new("/foo")),
@@ -1012,13 +1025,13 @@ mod tests {
         );
 
         assert_eq!(
-            fs.create_dir(&temp.path().join("foo")),
+            fs.create_dir(Path::new("foo")),
             Ok(()),
             "creating a directory",
         );
 
         assert_eq!(
-            fs.create_dir(&temp.path().join("foo/bar")),
+            fs.create_dir(Path::new("foo/bar")),
             Ok(()),
             "creating a sub-directory",
         );
@@ -1026,24 +1039,24 @@ mod tests {
         assert!(temp.path().join("foo/bar").exists(), "./foo/bar exists");
 
         assert_eq!(
-            fs.remove_dir(&temp.path().join("foo")),
+            fs.remove_dir(Path::new("foo")),
             Err(FsError::DirectoryNotEmpty),
             "removing a directory that has children",
         );
 
         assert_eq!(
-            fs.remove_dir(&temp.path().join("foo/bar")),
+            fs.remove_dir(Path::new("foo/bar")),
             Ok(()),
             "removing a sub-directory",
         );
 
         assert_eq!(
-            fs.remove_dir(&temp.path().join("foo")),
+            fs.remove_dir(Path::new("foo")),
             Ok(()),
             "removing a directory",
         );
 
-        let cur_dir = read_dir_names(&fs, temp.path());
+        let cur_dir = read_dir_names(&fs, "/");
 
         assert!(
             !cur_dir.contains(&"foo".to_string()),
@@ -1060,11 +1073,13 @@ mod tests {
 
     #[tokio::test]
     async fn test_rename() {
-        let temp = TempDir::new().unwrap();
-        let fs = FileSystem::default();
+        let temp: TempDir = TempDir::new().unwrap();
+        let fs = FileSystem::new(temp.path().into()).expect("get filesystem");
         std::fs::create_dir_all(temp.path().join("foo").join("qux")).unwrap();
-        let foo = temp.path().join("foo");
-        let bar = temp.path().join("bar");
+        let foo = Path::new("foo");
+        let bar = Path::new("bar");
+        let foo_realpath = temp.path().join(foo);
+        let bar_realpath = temp.path().join(bar);
 
         assert_eq!(
             fs.rename(Path::new("/"), Path::new("/bar")).await,
@@ -1114,7 +1129,7 @@ mod tests {
             "creating a new file (`hello2.txt`)",
         );
 
-        let cur_dir = read_dir_names(&fs, temp.path());
+        let cur_dir = read_dir_names(&fs, Path::new("/"));
 
         assert!(
             !cur_dir.contains(&"foo".to_string()),
@@ -1137,12 +1152,12 @@ mod tests {
         assert!(qux_dir.is_empty(), "the qux directory is empty");
 
         assert!(
-            bar.join("hello1.txt").exists(),
+            bar_realpath.join("hello1.txt").exists(),
             "the /bar/hello1.txt file exists"
         );
 
         assert!(
-            bar.join("hello2.txt").exists(),
+            bar_realpath.join("hello2.txt").exists(),
             "the /bar/hello2.txt file exists"
         );
 
@@ -1168,21 +1183,27 @@ mod tests {
             "renaming a file (in the same directory)",
         );
 
-        assert!(bar.exists(), "./bar exists");
-        assert!(bar.join("baz").exists(), "./bar/baz exists");
-        assert!(!foo.exists(), "foo does not exist anymore");
+        assert!(bar_realpath.exists(), "./bar exists");
+        assert!(bar_realpath.join("baz").exists(), "./bar/baz exists");
+        assert!(!foo_realpath.exists(), "foo does not exist anymore");
         assert!(
-            bar.join("baz/world2.txt").exists(),
+            bar_realpath.join("baz/world2.txt").exists(),
             "/bar/baz/world2.txt exists"
         );
         assert!(
-            bar.join("world1.txt").exists(),
+            bar_realpath.join("world1.txt").exists(),
             "/bar/world1.txt (ex hello1.txt) exists"
         );
-        assert!(!bar.join("hello1.txt").exists(), "hello1.txt was moved");
-        assert!(!bar.join("hello2.txt").exists(), "hello2.txt was moved");
         assert!(
-            bar.join("baz/world2.txt").exists(),
+            !bar_realpath.join("hello1.txt").exists(),
+            "hello1.txt was moved"
+        );
+        assert!(
+            !bar_realpath.join("hello2.txt").exists(),
+            "hello2.txt was moved"
+        );
+        assert!(
+            bar_realpath.join("baz/world2.txt").exists(),
             "world2.txt was moved to the correct place"
         );
     }
@@ -1194,9 +1215,9 @@ mod tests {
 
         let temp = TempDir::new().unwrap();
 
-        let fs = FileSystem::default();
+        let fs = FileSystem::new(temp.path().into()).expect("get filesystem");
 
-        let root_metadata = fs.metadata(temp.path()).unwrap();
+        let root_metadata = fs.metadata(Path::new("/")).unwrap();
 
         assert!(root_metadata.ft.dir);
         // it seems created is not evailable on musl, at least on CI testing.
@@ -1206,7 +1227,8 @@ mod tests {
         assert_eq!(root_metadata.modified, root_metadata.created);
         assert!(root_metadata.modified > 0);
 
-        let foo = temp.path().join("foo");
+        let foo = Path::new("foo");
+        let foo_realpath = temp.path().join(foo);
 
         assert_eq!(fs.create_dir(&foo), Ok(()));
 
@@ -1223,7 +1245,8 @@ mod tests {
 
         sleep(Duration::from_secs(3));
 
-        let bar = temp.path().join("bar");
+        let bar = Path::new("bar");
+        let bar_realpath = temp.path().join(bar);
 
         assert_eq!(fs.rename(&foo, &bar).await, Ok(()));
 
@@ -1242,26 +1265,26 @@ mod tests {
 
     #[tokio::test]
     async fn test_remove_file() {
-        let fs = FileSystem::default();
         let temp = TempDir::new().unwrap();
+        let fs = FileSystem::new(temp.path().into()).expect("get filesystem");
 
         assert!(
             matches!(
                 fs.new_open_options()
                     .write(true)
                     .create_new(true)
-                    .open(temp.path().join("foo.txt")),
+                    .open(Path::new("foo.txt")),
                 Ok(_)
             ),
             "creating a new file",
         );
 
-        assert!(read_dir_names(&fs, temp.path()).contains(&"foo.txt".to_string()));
+        assert!(read_dir_names(&fs, Path::new("/")).contains(&"foo.txt".to_string()));
 
         assert!(temp.path().join("foo.txt").is_file());
 
         assert_eq!(
-            fs.remove_file(&temp.path().join("foo.txt")),
+            fs.remove_file(&Path::new("foo.txt")),
             Ok(()),
             "removing a file that exists",
         );
@@ -1269,7 +1292,7 @@ mod tests {
         assert!(!temp.path().join("foo.txt").exists());
 
         assert_eq!(
-            fs.remove_file(&temp.path().join("foo.txt")),
+            fs.remove_file(&Path::new("foo.txt")),
             Err(FsError::EntryNotFound),
             "removing a file that doesn't exists",
         );
@@ -1278,34 +1301,22 @@ mod tests {
     #[tokio::test]
     async fn test_readdir() {
         let temp = TempDir::new().unwrap();
-        let fs = FileSystem::default();
+        let fs = FileSystem::new(temp.path().into()).expect("get filesystem");
 
+        assert_eq!(fs.create_dir(&Path::new("foo")), Ok(()), "creating `foo`");
         assert_eq!(
-            fs.create_dir(&temp.path().join("foo")),
-            Ok(()),
-            "creating `foo`"
-        );
-        assert_eq!(
-            fs.create_dir(&temp.path().join("foo/sub")),
+            fs.create_dir(&Path::new("foo/sub")),
             Ok(()),
             "creating `sub`"
         );
-        assert_eq!(
-            fs.create_dir(&temp.path().join("bar")),
-            Ok(()),
-            "creating `bar`"
-        );
-        assert_eq!(
-            fs.create_dir(&temp.path().join("baz")),
-            Ok(()),
-            "creating `bar`"
-        );
+        assert_eq!(fs.create_dir(&Path::new("bar")), Ok(()), "creating `bar`");
+        assert_eq!(fs.create_dir(&Path::new("baz")), Ok(()), "creating `bar`");
         assert!(
             matches!(
                 fs.new_open_options()
                     .write(true)
                     .create_new(true)
-                    .open(temp.path().join("a.txt")),
+                    .open(Path::new("a.txt")),
                 Ok(_)
             ),
             "creating `a.txt`",
@@ -1315,109 +1326,44 @@ mod tests {
                 fs.new_open_options()
                     .write(true)
                     .create_new(true)
-                    .open(&temp.path().join("b.txt")),
+                    .open(&Path::new("b.txt")),
                 Ok(_)
             ),
             "creating `b.txt`",
         );
 
-        let readdir = fs.read_dir(temp.path());
+        let readdir = fs.read_dir(Path::new("/"));
 
         assert!(
             readdir.is_ok(),
             "reading the directory `{}`",
-            temp.path().display()
+            Path::new("/").display()
         );
 
         let mut readdir = readdir.unwrap();
 
         let next = readdir.next().unwrap().unwrap();
         assert!(next.path.ends_with("a.txt"), "checking entry #1");
-        assert!(next.path.is_file(), "checking entry #1");
+        assert!(next.metadata().unwrap().is_file(), "checking entry #1");
 
         let next = readdir.next().unwrap().unwrap();
         assert!(next.path.ends_with("b.txt"), "checking entry #2");
-        assert!(next.path.is_file(), "checking entry #2");
+        assert!(next.metadata().unwrap().is_file(), "checking entry #2");
 
         let next = readdir.next().unwrap().unwrap();
         assert!(next.path.ends_with("bar"), "checking entry #3");
-        assert!(next.path.is_dir(), "checking entry #3");
+        assert!(next.metadata().unwrap().is_dir(), "checking entry #3");
 
         let next = readdir.next().unwrap().unwrap();
         assert!(next.path.ends_with("baz"), "checking entry #4");
-        assert!(next.path.is_dir(), "checking entry #4");
+        assert!(next.metadata().unwrap().is_dir(), "checking entry #4");
 
         let next = readdir.next().unwrap().unwrap();
         assert!(next.path.ends_with("foo"), "checking entry #5");
-        assert!(next.path.is_dir(), "checking entry #5");
+        assert!(next.metadata().unwrap().is_dir(), "checking entry #5");
 
         if let Some(s) = readdir.next() {
             panic!("next: {:?}", s);
         }
-    }
-
-    #[tokio::test]
-    async fn test_canonicalize() {
-        let temp = TempDir::new().unwrap();
-        std::fs::create_dir_all(temp.path().join("foo/bar/baz/qux")).unwrap();
-        std::fs::write(temp.path().join("foo/bar/baz/qux/hello.txt"), b"").unwrap();
-
-        let fs = FileSystem::default();
-        let root_dir = temp.path().canonicalize().unwrap();
-
-        assert_eq!(
-            fs.canonicalize(temp.path()),
-            Ok(root_dir.clone()),
-            "canonicalizing `/`",
-        );
-        assert_eq!(
-            fs.canonicalize(Path::new("foo")),
-            Err(FsError::InvalidInput),
-            "canonicalizing `foo`",
-        );
-        assert_eq!(
-            fs.canonicalize(&temp.path().join("././././foo/")),
-            Ok(root_dir.join("foo")),
-            "canonicalizing `/././././foo/`",
-        );
-        assert_eq!(
-            fs.canonicalize(&temp.path().join("foo/bar//")),
-            Ok(root_dir.join("foo").join("bar")),
-            "canonicalizing `/foo/bar//`",
-        );
-        assert_eq!(
-            fs.canonicalize(&temp.path().join("foo/bar/../bar")),
-            Ok(root_dir.join("foo").join("bar")),
-            "canonicalizing `/foo/bar/../bar`",
-        );
-        assert_eq!(
-            fs.canonicalize(&temp.path().join("foo/bar/../..")),
-            Ok(root_dir.clone()),
-            "canonicalizing `/foo/bar/../..`",
-        );
-        // temp.path().join("/foo/bar/../../..").exists() gives true on windows
-        #[cfg(not(target_os = "windows"))]
-        assert_eq!(
-            fs.canonicalize(&root_dir.join("/foo/bar/../../..")),
-            Err(FsError::InvalidInput),
-            "canonicalizing `/foo/bar/../../..`",
-        );
-        assert_eq!(
-            fs.canonicalize(&root_dir.join("C:/foo/")),
-            Err(FsError::InvalidInput),
-            "canonicalizing `C:/foo/`",
-        );
-        assert_eq!(
-            fs.canonicalize(&root_dir.join(
-                "foo/./../foo/bar/../../foo/bar/./baz/./../baz/qux/../../baz/./qux/hello.txt"
-            )),
-            Ok(root_dir
-                .join("foo")
-                .join("bar")
-                .join("baz")
-                .join("qux")
-                .join("hello.txt")),
-            "canonicalizing a crazily stupid path name",
-        );
     }
 }
