+++ lib/virtual-fs/src/fs/mod.rs
@@ -3,8 +3,6 @@ pub mod arc;
 pub mod empty;
 #[cfg(feature = "host-fs")]
 pub mod host;
-#[cfg(feature = "native-fs")]
-pub mod native;
 pub mod overlay;
 pub mod passthru;
 pub mod tmp;
